requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
redis==5.0.1
celery==5.3.6
chromadb==0.4.22
pyyaml==6.0.1
structlog==24.1.0
python-multipart==0.0.9
